Fixed sound issues.
The sum of all three PSG channels plus the buzzer did
sometimes overflow. I've changed the sign of channels B
and C, so that they subtract instead of adding. The result
is the sum of the inverted versions of the waveforms for channels
B and C. This is not noticeable to the human ear, but it keeps the
signal from overflowing.

Also, I've put the signal generation at the end of the PSG loop, once
the volumes are updated.