Some fixes (?) for Nouveau driver.
After changing nVidia-340 to Nouveau driver, I've found SpecIDE
crashing on startup. My guess is that the problem is rather in the
driver than in the program, but I've tried to fix the emulator
nevertheless.

- I've changed the vector<uint32> instances in the ULA, and used static
arrays of uint32 instead.
- Also, some included headers.

I haven't quite found the reason why the crashes happen, but now they
seem to be gone. I'm uploading this, and I hope for other drivers to
still work.