Sound levels changed.
Instead of adding all sound sources together, doubling it and
subtracting 0x4000, which causes a constant DC value in the sound
output, I'm trying adding ULA sounds + COVOX and then subtracting the
PSG output, or tape output - PSG on CPC.
This way the sound value is zero when no sound is being output. This
avoids loud clicks when the emulator starts or exits, and keeps the
signals in range.